requirements.txt
flake8==2.4.0
mccabe==0.3
paho-mqtt==1.1
pep8==1.5.7
pycurl==7.19.5.1
pyflakes==0.8.1
